+++ packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/features/sensororientation/DeviceOrientationManager.java
@@ -10,10 +10,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.res.Configuration;
-import android.hardware.SensorManager;
-import android.provider.Settings;
 import android.view.Display;
-import android.view.OrientationEventListener;
 import android.view.Surface;
 import android.view.WindowManager;
 import androidx.annotation.NonNull;
@@ -35,7 +32,6 @@ public class DeviceOrientationManager {
   private final boolean isFrontFacing;
   private final int sensorOrientation;
   private PlatformChannel.DeviceOrientation lastOrientation;
-  private OrientationEventListener orientationEventListener;
   private BroadcastReceiver broadcastReceiver;
 
   /** Factory method to create a device orientation manager. */
@@ -63,7 +59,7 @@ public class DeviceOrientationManager {
    *
    * <p>When orientation information is updated the new orientation is send to the client using the
    * {@link DartMessenger}. This latest value can also be retrieved through the {@link
-   * #getMediaOrientation()} accessor.
+   * #getVideoOrientation()} accessor.
    *
    * <p>If the device's ACCELEROMETER_ROTATION setting is enabled the {@link
    * DeviceOrientationManager} will report orientation updates based on the sensor information. If
@@ -71,55 +67,106 @@ public class DeviceOrientationManager {
    * the deliver orientation updates based on the UI orientation.
    */
   public void start() {
-    startSensorListener();
-    startUIListener();
+    if (broadcastReceiver != null) {
+      return;
+    }
+    broadcastReceiver =
+        new BroadcastReceiver() {
+          @Override
+          public void onReceive(Context context, Intent intent) {
+            handleUIOrientationChange();
+          }
+        };
+    activity.registerReceiver(broadcastReceiver, orientationIntentFilter);
+    broadcastReceiver.onReceive(activity, null);
   }
 
   /** Stops listening for orientation updates. */
   public void stop() {
-    stopSensorListener();
-    stopUIListener();
+    if (broadcastReceiver == null) {
+      return;
+    }
+    activity.unregisterReceiver(broadcastReceiver);
+    broadcastReceiver = null;
   }
 
   /**
-   * Returns the last captured orientation in degrees based on sensor or UI information.
+   * Returns the device's photo orientation in degrees based on the sensor orientation and the last
+   * known UI orientation.
    *
-   * <p>The orientation is returned in degrees and could be one of the following values:
+   * <p>Returns one of 0, 90, 180 or 270.
    *
-   * <ul>
-   *   <li>0: Indicates the device is currently in portrait.
-   *   <li>90: Indicates the device is currently in landscape left.
-   *   <li>180: Indicates the device is currently in portrait down.
-   *   <li>270: Indicates the device is currently in landscape right.
-   * </ul>
-   *
-   * @return The last captured orientation in degrees
+   * @return The device's photo orientation in degrees.
    */
-  public int getMediaOrientation() {
-    return this.getMediaOrientation(this.lastOrientation);
+  public int getPhotoOrientation() {
+    return this.getPhotoOrientation(this.lastOrientation);
   }
 
   /**
-   * Returns the device's orientation in degrees based on the supplied {@link
-   * PlatformChannel.DeviceOrientation} value.
+   * Returns the device's photo orientation in degrees based on the sensor orientation and the
+   * supplied {@link PlatformChannel.DeviceOrientation} value.
    *
-   * <p>
-   *
-   * <ul>
-   *   <li>PORTRAIT_UP: converts to 0 degrees.
-   *   <li>LANDSCAPE_LEFT: converts to 90 degrees.
-   *   <li>PORTRAIT_DOWN: converts to 180 degrees.
-   *   <li>LANDSCAPE_RIGHT: converts to 270 degrees.
-   * </ul>
+   * <p>Returns one of 0, 90, 180 or 270.
    *
    * @param orientation The {@link PlatformChannel.DeviceOrientation} value that is to be converted
    *     into degrees.
-   * @return The device's orientation in degrees.
+   * @return The device's photo orientation in degrees.
    */
-  public int getMediaOrientation(PlatformChannel.DeviceOrientation orientation) {
+  public int getPhotoOrientation(PlatformChannel.DeviceOrientation orientation) {
+    int angle = 0;
+    // Fallback to device orientation when the orientation value is null.
+    if (orientation == null) {
+      orientation = getUIOrientation();
+    }
+
+    switch (orientation) {
+      case PORTRAIT_UP:
+        angle = 90;
+        break;
+      case PORTRAIT_DOWN:
+        angle = 270;
+        break;
+      case LANDSCAPE_LEFT:
+        angle = isFrontFacing ? 180 : 0;
+        break;
+      case LANDSCAPE_RIGHT:
+        angle = isFrontFacing ? 0 : 180;
+        break;
+    }
+
+    // Sensor orientation is 90 for most devices, or 270 for some devices (eg. Nexus 5X).
+    // This has to be taken into account so the JPEG is rotated properly.
+    // For devices with orientation of 90, this simply returns the mapping from ORIENTATIONS.
+    // For devices with orientation of 270, the JPEG is rotated 180 degrees instead.
+    return (angle + sensorOrientation + 270) % 360;
+  }
+
+  /**
+   * Returns the device's video orientation in degrees based on the sensor orientation and the last
+   * known UI orientation.
+   *
+   * <p>Returns one of 0, 90, 180 or 270.
+   *
+   * @return The device's video orientation in degrees.
+   */
+  public int getVideoOrientation() {
+    return this.getVideoOrientation(this.lastOrientation);
+  }
+
+  /**
+   * Returns the device's video orientation in degrees based on the sensor orientation and the
+   * supplied {@link PlatformChannel.DeviceOrientation} value.
+   *
+   * <p>Returns one of 0, 90, 180 or 270.
+   *
+   * @param orientation The {@link PlatformChannel.DeviceOrientation} value that is to be converted
+   *     into degrees.
+   * @return The device's video orientation in degrees.
+   */
+  public int getVideoOrientation(PlatformChannel.DeviceOrientation orientation) {
     int angle = 0;
 
-    // Fallback to device orientation when the orientation value is null
+    // Fallback to device orientation when the orientation value is null.
     if (orientation == null) {
       orientation = getUIOrientation();
     }
@@ -146,51 +193,9 @@ public class DeviceOrientationManager {
     return (angle + sensorOrientation + 360) % 360;
   }
 
-  private void startSensorListener() {
-    if (orientationEventListener != null) {
-      return;
-    }
-    orientationEventListener =
-        new OrientationEventListener(activity, SensorManager.SENSOR_DELAY_NORMAL) {
-          @Override
-          public void onOrientationChanged(int angle) {
-            handleSensorOrientationChange(angle);
-          }
-        };
-    if (orientationEventListener.canDetectOrientation()) {
-      orientationEventListener.enable();
-    }
-  }
-
-  private void startUIListener() {
-    if (broadcastReceiver != null) {
-      return;
-    }
-    broadcastReceiver =
-        new BroadcastReceiver() {
-          @Override
-          public void onReceive(Context context, Intent intent) {
-            handleUIOrientationChange();
-          }
-        };
-    activity.registerReceiver(broadcastReceiver, orientationIntentFilter);
-    broadcastReceiver.onReceive(activity, null);
-  }
-
-  /**
-   * Handles orientation changes based on information from the device's sensors.
-   *
-   * <p>This method is visible for testing purposes only and should never be used outside this
-   * class.
-   *
-   * @param angle of the current orientation.
-   */
-  @VisibleForTesting
-  void handleSensorOrientationChange(int angle) {
-    if (!isAccelerometerRotationLocked()) {
-      PlatformChannel.DeviceOrientation orientation = calculateSensorOrientation(angle);
-      lastOrientation = handleOrientationChange(orientation, lastOrientation, messenger);
-    }
+  /** @return the last received UI orientation. */
+  public PlatformChannel.DeviceOrientation getLastUIOrientation() {
+    return this.lastOrientation;
   }
 
   /**
@@ -201,10 +206,9 @@ public class DeviceOrientationManager {
    */
   @VisibleForTesting
   void handleUIOrientationChange() {
-    if (isAccelerometerRotationLocked()) {
-      PlatformChannel.DeviceOrientation orientation = getUIOrientation();
-      lastOrientation = handleOrientationChange(orientation, lastOrientation, messenger);
-    }
+    PlatformChannel.DeviceOrientation orientation = getUIOrientation();
+    handleOrientationChange(orientation, lastOrientation, messenger);
+    lastOrientation = orientation;
   }
 
   /**
@@ -215,37 +219,13 @@ public class DeviceOrientationManager {
    * class.
    */
   @VisibleForTesting
-  static DeviceOrientation handleOrientationChange(
+  static void handleOrientationChange(
       DeviceOrientation newOrientation,
       DeviceOrientation previousOrientation,
       DartMessenger messenger) {
     if (!newOrientation.equals(previousOrientation)) {
       messenger.sendDeviceOrientationChangeEvent(newOrientation);
     }
-
-    return newOrientation;
-  }
-
-  private void stopSensorListener() {
-    if (orientationEventListener == null) {
-      return;
-    }
-    orientationEventListener.disable();
-    orientationEventListener = null;
-  }
-
-  private void stopUIListener() {
-    if (broadcastReceiver == null) {
-      return;
-    }
-    activity.unregisterReceiver(broadcastReceiver);
-    broadcastReceiver = null;
-  }
-
-  private boolean isAccelerometerRotationLocked() {
-    return android.provider.Settings.System.getInt(
-            activity.getContentResolver(), Settings.System.ACCELEROMETER_ROTATION, 0)
-        != 1;
   }
 
   /**