+++ packages/connectivity/connectivity/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Remove references to the Android V1 embedding.
+* Updated Android lint settings.
+
 ## 3.0.6
 
 * Update README to point to Plus Plugins version.